refac: seperate post entity and repository from center
